Adds beginnings of equipping and unequipping items [IN PROGRESS]
Adds equipping and unequipping items, still being worked on now. As of now, items are not restored to the backpack upon equipping a different item (will fix soon). Also casting answer to lowercase for inclusivity, and adding the message if multiple items match the partial name. Work more on it tomorrow.