Finish equip/unequip, adds equipped items screen
Finishes equipping and unequipping mechanics and commands using -equip and unequip. Also adds equipped items screen with -equipped, until we make it a second page on the stats screen. Next most important thing is to add a second screen on the backpack as if you have too much loot in there it won't display because it goes over 2000 characters. Need to get a check for 1500 characters and remainder and create a second page. Also to make the blank spaces not show up if the slots aren't populated. Also actual item mechanics we talked about with the rarity number, random chance to have prefixes/suffixes etc, as well as sale price. This commit is the basic framework for all of that; officially good to pull in.